+++ probpipe/modeling/_likelihood.py
@@ -222,7 +222,7 @@ class IncrementalConditioner[P, D](Module):
                 raise ValueError("Cannot provide both `data` and named data kwargs")
             from ..core.record import Record
 
-            data = Record("data", kwargs, name_is_auto=True)
+            data = Record("data", kwargs)
         posterior = self._step(self._curr_posterior, data)
         self._curr_posterior = posterior
         return posterior